+++ src/Button/Button.tsx
@@ -34,7 +34,7 @@ interface ButtonProps
     size?: ButtonSize;
 }
 
-const Button: React.FC<ButtonProps> = ({
+const Button: React.FC<React.PropsWithChildren<ButtonProps>> = ({
     children,
     className,
     variant,